# Speedup `DFSchema::merge` using HashSet indices