@matthewmturner 2024-01-27
Nice 😀 

Trade off seems worth it. 

I'm curious what's going on in that TPCH benchmark though as the other physical select benchmark showed solid improvement so not sure it's all physical planning.