@alamb 2024-01-29
> Speeds up benchmarks in sql_planner.rs by 18-75%. However, physical planning benchmarks seem to regress by 1.8%. I wonder if that is an acceptable trade?



I think this is an acceptable trade